Fix metrics management in pipelines
Previously, metrics computed were hardcoded instead of using the defaults or
the ones given by the configuration; moreover, the configuration was specified
only for benchmark, and actually it did not make sense, as benchmark only uses
a single metric to select the best model. It is in the training process of a
model that multiple metrics are reported.

Therefore, the commit includes the following:
- Use of metrics in training pipeline
- Specific default metrics for each problem type
- Possibility of specify metrics in configuration file
- Usage of both sklearn metrics and custom ones