invalidate_experiment for benchmark + torch implementation
This commit includes a generic feature (invalidate_experiment) which is
implemented for torch model benchmarking. The idea is that there could be
combinations of parameters that do not make sense, and as it is quite complex
to exclude scenarios in advance, they are invalidated within the process but
before cross-validating models.
Technical changes are the following:
- Inclusion of invalidate_experiment method within BaseBenchmark class, and an
  implementation of this method in TorchBenchmark class
- Call to the aforementioned method within EstimatorTrainer class, and
  previously pass required parameters for this
- Corresponding tests
- Clean a couple of method specifications about benchmarks