Rewrite examples in order to work outside development
Before, examples only worked if within development environment, but not if the
library was directly installed (via pip).
The way to handle this is the following:
- Place example data in *honcaml/config/examples/data*
- Place example configuration in *honcaml/config/examples/files*
- Add a command line option that retrieves internal data and configuration
  files in order to place them in the specified directory
When switching to the specified directory, it is possible to run any pipeline
within the *files* directory, which correctly refers to their respective data.
Moreover, configuration files have been renamed to a clearer name. Benchmark
configurations include both ML models and DL models.