Update max_features default search space
Previously,. default search space for hyperparamereter max_features, used in
both RandomForest and GradientBoosting models (for regression and
classification), considered value 1. This did not make sense, as if it were an
int, it should be treated as randint between values, but in this case for each
split, if was the chosen one, just a single feature was tested. As it is not
easy for a single hyperparameter to have a randint within a choice, by default
a string choice is left, which are both feasible options